Se valida que el id pasado por parametro sea un númerico

+++ app/controllers/PerfilRolController.mjs
@@ -1,5 +1,7 @@
 import {PerfilRol} from "../models/index.mjs";
 import HttpCode from "../../configs/httpCode.mjs";
+import UnprocessableEntityException from "../../handlers/UnprocessableEntityException.mjs";
+
 
 export default class PerfilController {
 
@@ -19,9 +21,13 @@ export default class PerfilController {
     }
 
     static async show(req, res) {
-        const perfil_rol = await PerfilRol.findOne({
+        const { id } = req.params; 
+        if(isNaN(id))
+            throw new UnprocessableEntityException("UNPROCESSABLE_ENTITY", 422, "El parámetro no es un id válido");
+        
+            const perfil_rol = await PerfilRol.findOne({
             where: {
-                id: req.params.id
+                id
             }
         })
         return res.status(HttpCode.HTTP_OK).json(perfil_rol);
@@ -44,9 +50,13 @@ export default class PerfilController {
     }
 
     static async destroy(req, res) {
+        const { id } = req.params; 
+        if(isNaN(id))
+            throw new UnprocessableEntityException("UNPROCESSABLE_ENTITY", 422, "El parámetro no es un id válido");
+
         await PerfilRol.destroy({
             where: {
-                id: req.params.id
+                id
             },
         })
         return res.status(HttpCode.HTTP_OK).json({
