+++ app/controllers/RolController.mjs
@@ -1,5 +1,7 @@
 import {Rol} from "../models/index.mjs";
 import HttpCode from "../../configs/httpCode.mjs";
+import UnprocessableEntityException from "../../handlers/UnprocessableEntityException.mjs";
+
 
 export default class RolController {
 
@@ -19,9 +21,13 @@ export default class RolController {
     }
 
     static async show(req, res) {
+        const { id } = req.params; 
+        if(isNaN(id))
+            throw new UnprocessableEntityException("UNPROCESSABLE_ENTITY", 422, "El parámetro no es un id válido");
+
         const rol = await Rol.findOne({
             where: {
-                id: req.params.id
+                id
             }
         })
 
@@ -31,11 +37,14 @@ export default class RolController {
 
     static async update(req, res) {
         const {name} = req.body
+        const { id } = req.params; 
+        if(isNaN(id))
+            throw new UnprocessableEntityException("UNPROCESSABLE_ENTITY", 422, "El parametro no es un id válido");
         const rol = await Rol.update({
             name
         }, {
             where: {
-                id: req.params.id
+                id
             },
             returning: ['name']
         })
@@ -43,9 +52,12 @@ export default class RolController {
     }
 
     static async destroy(req, res) {
+        const { id } = req.params; 
+        if(isNaN(id))
+            throw new UnprocessableEntityException("UNPROCESSABLE_ENTITY", 422, "El parametro no es un id válido");
         await Rol.destroy({
             where: {
-                id: req.params.id
+                id
             },
         })
